+++ CHANGELOG.md
@@ -25,8 +25,14 @@
 - Sample was improperly setup
   - Now the sample can be imported
 
-##  [1.1.2] - 16. April 2025
+## [1.1.2] - 16. April 2025
 ### Bug Fixing
 
 - Fixed a bug where pause menu wouldn't close, and wouldn't be on top
 - Pause menu script namespace
+
+## [1.2.2] - 16. April 2025
+### Scene Management
+
+- Added a loading screen
+  - Therefore, added a scene loader script to manage scene loading